# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "Smm Api Documentation",
        "contact": {
            "email": "takahiroasiro@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin/User"
                ],
                "summary": "Find user by name or email",
                "description": "Администратор: поиск пользователя по имени или email",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name parameter in query",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email parameter in query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/admin/find": {
            "get": {
                "tags": [
                    "Admin/User"
                ],
                "summary": "Find user by name or email with balances",
                "description": "Администратор: поиск пользователя по имени или emailс их балансами",
                "operationId": "find",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name parameter in query",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email parameter in query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/articles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Articles list",
                "description": "Получение списка статей",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "The offset parameter in query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The limit parameter in query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "The sort parameter in query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "Creating an article",
                "description": "Создание статьи",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "The slug parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cover",
                        "in": "query",
                        "description": "The cover parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "heading",
                        "in": "query",
                        "description": "The heading parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "The description parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "The tags parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "article",
                        "in": "query",
                        "description": "The article parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "headtitle",
                        "in": "query",
                        "description": "The headtitle parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "headdescription",
                        "in": "query",
                        "description": "The headdescription parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article created"
                    }
                }
            }
        },
        "/api/articles/{slug}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Getting a unique article",
                "description": "Получение уникальной статьи",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "The offset parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/articles/{id}": {
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "update articles",
                "description": "Обновление статей",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "request",
                        "in": "query",
                        "description": "The request object parameter in query",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "destroy articles",
                "description": "Удаление статей",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The request id parameter in query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles deleted successfully"
                    }
                }
            }
        },
        "/api/login/local": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Local login",
                "description": "Локальный логин: login/email + password",
                "operationId": "localLogin",
                "parameters": [
                    {
                        "name": "login",
                        "in": "query",
                        "description": "The login parameter in query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email parameter in query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password parameter in query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/update_avatar": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update avatar",
                "description": "Обновить аватар",
                "operationId": "updateAvatar",
                "parameters": [
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "The avatar parameter mimes: jpeg, jpg, png",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "avatar changed"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}